Consolidating hashing code. (#1207)
